Reorganize Wear OS Material3 list composables
- Move list-related composables to dedicated 'list' package
- Split `OssLicenseListScreen.kt` into focused components
- Remove redundant `OssLicenseListImpl` layer
