+++ ui-wear-compose-material3/src/main/java/com/github/droibit/oss_licenses/ui/wear/compose/material3/internal/OssLicenseNavGraph.kt
@@ -13,6 +13,7 @@ import androidx.wear.compose.navigation.SwipeDismissableNavHost
 import androidx.wear.compose.navigation.composable
 import androidx.wear.compose.navigation.rememberSwipeDismissableNavController
 import com.github.droibit.oss_licenses.ui.viewmodel.OssLicenseViewModel
+import com.github.droibit.oss_licenses.ui.wear.compose.material3.internal.list.OssLicenseListScreen
 import com.github.droibit.oss_licenses.ui.wear.compose.navigation.Routes.LicenseDetail
 import com.github.droibit.oss_licenses.ui.wear.compose.navigation.Routes.LicenseList
 import com.github.droibit.oss_licenses.ui.wear.compose.navigation.navigateToDetail
