+++ ui-wear-compose-material3/src/main/java/com/github/droibit/oss_licenses/ui/wear/compose/material3/internal/list/OssLicenseList.kt
@@ -1,9 +1,8 @@
-package com.github.droibit.oss_licenses.ui.wear.compose.material3.internal
+package com.github.droibit.oss_licenses.ui.wear.compose.material3.internal.list
 
 import androidx.annotation.RestrictTo
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.wear.compose.foundation.ExperimentalWearFoundationApi
 import androidx.wear.compose.foundation.lazy.TransformingLazyColumn
 import androidx.wear.compose.foundation.lazy.TransformingLazyColumnItemScope
 import androidx.wear.compose.foundation.lazy.TransformingLazyColumnState
@@ -13,7 +12,6 @@ import com.github.droibit.oss_licenses.parser.OssLicense
 import com.github.droibit.oss_licenses.ui.wear.compose.core.rememberResponsiveColumnPadding
 
 @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
-@OptIn(ExperimentalWearFoundationApi::class)
 @Composable
 fun OssLicenseList(
   licenses: List<OssLicense>,
