+++ ui-wear-compose-material3/src/test/java/com/github/droibit/oss_licenses/ui/wear/compose/material3/internal/list/OssLicenseListScreenshotTest.kt
@@ -1,4 +1,4 @@
-package com.github.droibit.oss_licenses.ui.wear.compose.material3.internal
+package com.github.droibit.oss_licenses.ui.wear.compose.material3.internal.list
 
 import androidx.wear.compose.foundation.lazy.TransformingLazyColumnState
 import com.github.droibit.oss_licenses.ui.compose.screenshots.TestFixtures
